change data fetching filter component
